@@ -1,11 +1,18 @@
 module github.com/godoes/md2html/v2
 
-go 1.15
+go 1.17
 
 require (
-	github.com/PuerkitoBio/goquery v1.6.0
-	github.com/alecthomas/chroma v0.8.3-0.20201223195406-5b9087ac1446
-	github.com/jessevdk/go-flags v1.4.0
+	github.com/PuerkitoBio/goquery v1.8.0
+	github.com/alecthomas/chroma v0.10.0
+	github.com/jessevdk/go-flags v1.5.0
 	github.com/nocd5/goldmark-highlighting v0.0.0-20201224144305-f7ed7f755f7e
-	github.com/yuin/goldmark v1.3.0
+	github.com/yuin/goldmark v1.5.3
+)
+
+require (
+	github.com/andybalholm/cascadia v1.3.1 // indirect
+	github.com/dlclark/regexp2 v1.8.0 // indirect
+	golang.org/x/net v0.5.0 // indirect
+	golang.org/x/sys v0.4.0 // indirect
 )
